Add files via upload
Sample codes of the algorithm proposed in my paper with the subject "Designing a scheduling decision support system for the skin pass line: A case study of the steel finishing line". 
I have developed a commercial decision support system for the implementation of the proposed algorithm in the steel company. Therefore, input and output of code link to the database of the company.